+++ conreality_headset/lib/src/headset_event.dart
@@ -1,0 +1,4 @@
+/* This is free and unencumbered software released into the public domain. */
+
+/// Represents a headset device event.
+abstract class HeadsetEvent {}