+++ conreality_headset/lib/src/headset_type.dart
@@ -1,0 +1,8 @@
+/* This is free and unencumbered software released into the public domain. */
+
+/// Represents the headset device type (wireless or wired).
+enum HeadsetType {
+  unknown,
+  wired, // headphone jack plug
+  wireless, // Bluetooth
+}